Width of plots can be specified in LaTeX document
Modified the template using some pgfkeys magic to check for the existence
of the /artist/width key.  If it exists, use its value to define the width
of the plot.  If it doesn't, use the value specified in the python source.

In effect, the width of the plot can be chosen at the time of creation of
the plot (in python), but can be overridden at the time of publication (in
LaTeX).