Add colorbar for MultiPlot
Currently adds colorbar to last plot.
This is always in the lower right corner.
By default the colorbar is to the right of a plot.
The horizontal colorbar is below plots.
So this is safe, until/unless the colobar can be to the top or left.
Colorbar is not yet scaled to the full height/width of all plots