Limit data to x-axis limits
Limit the data points to data points inside the x-axis limits. This is especially useful when plotting a very large dataset while limiting the view to a small range. Do not include the full dataset in the TeX-files, because TeX will take a very long time to plot the graph and probably run out of memory.